+++ src/Plivo/Resource/Application/ApplicationInterface.cs
@@ -54,6 +54,7 @@ namespace Plivo.Resource.Application
         {
             var mandatoryParams = new List<string> { "appName" }
                 ;
+            bool isVoiceRequest = true;
             var data = CreateData(
                 mandatoryParams,
                 new
@@ -70,7 +71,8 @@ namespace Plivo.Resource.Application
                     defaultNumberApp,
                     defaultEndpointApp,
                     subaccount,
-                    logIncomingMessages
+                    logIncomingMessages,
+                    isVoiceRequest
                 });
 
 			return ExecuteWithExceptionUnwrap(() =>
@@ -109,6 +111,7 @@ namespace Plivo.Resource.Application
         {
             var mandatoryParams = new List<string> {"appName"}
                 ;
+            bool isVoiceRequest = true;
             var data = CreateData(
                 mandatoryParams,
                 new
@@ -125,7 +128,8 @@ namespace Plivo.Resource.Application
                     defaultNumberApp,
                     defaultEndpointApp,
                     subaccount,
-                    logIncomingMessages
+                    logIncomingMessages,
+                    isVoiceRequest
                 });
 
             var result = await Client.Update<ApplicationCreateResponse>(Uri, data);
@@ -144,7 +148,7 @@ namespace Plivo.Resource.Application
         {
 			return ExecuteWithExceptionUnwrap(() =>
 			{
-				var application = Task.Run(async () => await GetResource<Application>(appId).ConfigureAwait(false)).Result;
+				var application = Task.Run(async () => await GetResource<Application>(appId, new Dictionary<string, object> () { {"is_voice_request", true} }).ConfigureAwait(false)).Result;
 				application.Interface = this;
 				return application;
 			});
@@ -156,7 +160,7 @@ namespace Plivo.Resource.Application
         /// <param name="appId">App identifier.</param>
         public async Task<Application> GetAsync(string appId)
         {
-            var application = await GetResource<Application>(appId);
+            var application = await GetResource<Application>(appId, new Dictionary<string, object> () { {"is_voice_request", true} });
             application.Interface = this;
             return application;
         }
@@ -174,8 +178,9 @@ namespace Plivo.Resource.Application
             string subaccount = null, uint? limit = null, uint? offset = null)
         {
             var mandatoryParams = new List<string> { "" };
+            bool isVoiceRequest = true;
             var data = CreateData(
-                mandatoryParams, new { subaccount, limit, offset });
+                mandatoryParams, new { subaccount, limit, offset, isVoiceRequest });
 
 			return ExecuteWithExceptionUnwrap(() =>
 			{
@@ -198,8 +203,9 @@ namespace Plivo.Resource.Application
             string subaccount = null, uint? limit = null, uint? offset = null)
         {
             var mandatoryParams = new List<string> {""};
+            bool isVoiceRequest = true;
             var data = CreateData(
-                mandatoryParams, new {subaccount, limit, offset});
+                mandatoryParams, new {subaccount, limit, offset, isVoiceRequest});
 
             var resources = await ListResources<ListResponse<Application>>(data);
             resources.Objects.ForEach(
@@ -216,11 +222,17 @@ namespace Plivo.Resource.Application
         /// </summary>
         /// <returns>The delete.</returns>
         /// <param name="appId">App identifier.</param>
-        public DeleteResponse<Application> Delete(string appId)
+        /// <param name="cascade">Cascade.</param>
+        /// <param name="newEndpointApplication">New Endpoint Application.</param>
+        public DeleteResponse<Application> Delete(string appId, bool? cascade = null, string newEndpointApplication = null)
         {
+            var data = new Dictionary<string, object> {};
+            bool isVoiceRequest = true;
+            data = CreateData(new List<string> {}, new {cascade, newEndpointApplication, isVoiceRequest});
+
 			return ExecuteWithExceptionUnwrap(() =>
 			{
-				return Task.Run(async () => await DeleteResource<DeleteResponse<Application>>(appId).ConfigureAwait(false)).Result;
+				return Task.Run(async () => await DeleteResource<DeleteResponse<Application>>(appId, data).ConfigureAwait(false)).Result;
 			});
 		}
         /// <summary>
@@ -228,9 +240,14 @@ namespace Plivo.Resource.Application
         /// </summary>
         /// <returns>The delete.</returns>
         /// <param name="appId">App identifier.</param>
-        public async Task<DeleteResponse<Application>> DeleteAsync(string appId)
+        /// <param name="cascade">Cascade.</param>
+        /// <param name="newEndpointApplication">New Endpoint Application.</param>
+        public async Task<DeleteResponse<Application>> DeleteAsync(string appId, bool? cascade = null, string newEndpointApplication = null)
         {
-            return await DeleteResource<DeleteResponse<Application>>(appId);
+            var data = new Dictionary<string, object> { };
+            bool isVoiceRequest = true;
+            data = CreateData(new List<string> { }, new { cascade, newEndpointApplication, isVoiceRequest });
+            return await DeleteResource<DeleteResponse<Application>>(appId, data);
         }
         #endregion
 
@@ -262,6 +279,7 @@ namespace Plivo.Resource.Application
             string subaccount = null, bool? logIncomingMessages = null)
         {
             var mandatoryParams = new List<string> { "" };
+            bool isVoiceRequest = true;
             var data = CreateData(
                 mandatoryParams,
                 new
@@ -277,7 +295,8 @@ namespace Plivo.Resource.Application
                     defaultNumberApp,
                     defaultEndpointApp,
                     subaccount,
-                    logIncomingMessages
+                    logIncomingMessages,
+                    isVoiceRequest
                 });
 
 			return ExecuteWithExceptionUnwrap(() =>
@@ -314,6 +333,7 @@ namespace Plivo.Resource.Application
             string subaccount = null, bool? logIncomingMessages = null)
         {
             var mandatoryParams = new List<string> {""};
+            bool isVoiceRequest = true;
             var data = CreateData(
                 mandatoryParams,
                 new
@@ -329,7 +349,8 @@ namespace Plivo.Resource.Application
                     defaultNumberApp,
                     defaultEndpointApp,
                     subaccount,
-                    logIncomingMessages
+                    logIncomingMessages,
+                    isVoiceRequest
                 });
 
             var result = await Client.Update<UpdateResponse<Application>>(Uri + appId + "/", data);